ref(auth): Refactor send email verification
