+++ app/server/src/controllers/AuthController.ts
@@ -1,16 +1,13 @@
-import jwt from 'jsonwebtoken';
 import fs from 'fs';
 
 import { Request, Response } from "express";
-import { handleError, validatePayload, registerPayloadSchema, verifyEmailSchema } from "../common/http";
+import { InvariantError, PayloadError } from "../common/exception";
+import { handleError, registerPayloadSchema, validatePayload, verifyEmailSchema } from "../common/http";
+import { AuthService } from "../services";
 import { IInstitution, RegisterPayloadType } from "../types/auth";
-import { InvariantError, NotFoundError, PayloadError } from "../common/exception";
-import { EmailService, AuthService } from "../services";
-import { INodemailerMessage } from '../types/email';
-import { generateFutureDate } from '../common/utils';
 
 export class AuthController {
-    constructor(public authService: AuthService, public emailService: EmailService) {
+    constructor(public authService: AuthService) {
     }
 
     async register(req: Request, res: Response) {
@@ -97,35 +94,14 @@ export class AuthController {
         try {
             validatePayload(verifyEmailSchema, req.body);
             const { email } = req.body;
-            const { user } = await this.authService.getUserByUniqueIdentity(email);
-            const generatedToken = jwt.sign({
-                id: user.id,
-                email: user.email,
-                username: user.username,
-                role: user.role_id
-            }, process.env.SECRET_ACCESS_TOKEN ?? 'SECRET_TOKEN_RAHASIA', {
-                expiresIn: 3600 * 24 * 7
-            });
-            if (user.is_verified) {
+            const { user } = await this.authService.sendEmailVerification(email);
+
+            if (!user) {
                 return res.status(200).json({
                     status: 'Not Changed',
                     message: 'User Already Verified'
                 })
             }
-
-            const generatedDate = generateFutureDate(7);
-
-            const payload: INodemailerMessage = {
-                from: process.env.SMTP_EMAIL ?? 'admin@gmail.com',
-                to: email,
-                subject: 'Email Verification',
-                html: `
-                <h1>Email Verification</h1>
-                <p>Click this link before <strong>${generatedDate}</strong> to verify your email: <a href="${process.env.API_BASE_URL ?? 'http://localhost:5000'}/auth/email/verify?token=${generatedToken}">Verify Email</a></p>
-                `
-            }
-
-            await this.emailService.sendEmail(payload);
 
             res.status(200).json({
                 status: 'Success',
